pyproject.toml
[build-system]
requires = ["setuptools==75.1.0", "wheel==0.44.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jssp-yafs"
version = "0.1.0"
description = "Multi-objective JSSP optimization with YAFS-coupled edge/fog/cloud simulation"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = {text = "MIT"}
authors = [
  {name = "Research Engineering Template"}
]
dependencies = [
  "numpy==1.26.4",
  "pandas==2.2.2",
  "networkx==3.2.1",
  "matplotlib==3.10.8",
  "seaborn==0.13.2",
  "scipy==1.14.1",
  "pyyaml==6.0.2",
  "tqdm==4.66.5",
  "diskcache==5.6.3",
  "pymoo==0.6.1.3",
  "yafs @ git+https://github.com/acsicuib/YAFS.git@09777042bf50c38a1c3edaa2cbf32ab63c4d6824"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3"
]

[project.scripts]
jssp-yafs = "jssp_yafs.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = ["results", "data/processed"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E203"]
